https://frdm-tfc.googlecode.com/svn/trunk/RELEASES/R0.2Changed the crystal oscillator mode to use "low power" drive.  Some freedom boards would not work unless it was set for this.  The problem manifested itself was non function timer/PWM code.
The root of the problem is in the Cyrstal Oscillator/PLL.   Long story short,   *some* of the boards have crystals that will not start in "High Gain Mode".  In these cases, the oscillator needs to be configured for "LOW_POWER" mode.  High gain usually works for all crystals but whatever date code, production run units are on some of the boards do not like this.

The reason PWN was not working that all the timers were derived from the PLL / Crystal.   When it failed to start it falls back to a internal reference clock and the TPM modules no longer had a valid clock.

When I was help a group today we configure a pin used for a servo as GPIO.   It was then toggled in the SysTick IRQ route (1mS rate).  I  notice that the rate was NOT 1mSecs.   I traced it back to the weird crystals....